Apply checks on the number of point received to make sure we have enough to process.

+++ py/inundation_filter.py
@@ -43,6 +43,12 @@ def inundation(ins,outs):
     below_array = np_array[(elevation >= np_array[:,2]) & (np_array[:,2] >= 0)][:,0:3]
     above_array = np_array[(elevation < np_array[:,2]) & (np_array[:,2] < (elevation + 3))][:,0:3] # arbitrary top filter (watch elevation of location)
 
+    # check that we have points before trying to do anything with them
+    if len(below_array) < 1000:
+        sys.stdout.write("\nThere are %s points falling below the requested elevation. Try a higher elevation.\n\n"%(len(below_array)))
+        sys.stdout.flush()
+        return False
+
     # # below floodplain
     # below_points = gpd.GeoSeries([i for i in asMultiPoint(below_array) if not i.is_empty])
     below_points = gpd.GeoSeries([Point(i) for i in below_array])
@@ -66,35 +72,40 @@ def inundation(ins,outs):
     sys.stdout.write("Below polygon successfully exported\n")
     sys.stdout.flush()
 
+    # check that we have points before trying to do anything with them
+    if len(above_array) < 1000:
+        sys.stdout.write("\nThere are %s points above the requested elevation. Not processing the above point and just exporting area below requested elevation.\n\n"%(len(above_array)))
+        sys.stdout.flush()
+        inundation_gdf = below_polygon_gdf
+    else:
+        # above floodplain
+        above_points = gpd.GeoSeries([Point(i) for i in above_array])
+        above_gdf = gpd.GeoDataFrame(crs=crs,geometry=[i for i in above_points])
+        above_gdf.to_file('shp/%s_above_points.shp'%(outputFileName))
+        # fiona can't overwrite a geojson, remove first
+        removeFile("geojson/%s_above_points.geojson"%(outputFileName))
+        above_gdf.to_file('geojson/%s_above_points.geojson'%(outputFileName), driver='GeoJSON')
+        sys.stdout.write("Above points successfully exported\n")
+        sys.stdout.flush()
 
-    # above floodplain
-    above_points = gpd.GeoSeries([Point(i) for i in above_array])
-    above_gdf = gpd.GeoDataFrame(crs=crs,geometry=[i for i in above_points])
-    above_gdf.to_file('shp/%s_above_points.shp'%(outputFileName))
-    # fiona can't overwrite a geojson, remove first
-    removeFile("geojson/%s_above_points.geojson"%(outputFileName))
-    above_gdf.to_file('geojson/%s_above_points.geojson'%(outputFileName), driver='GeoJSON')
-    sys.stdout.write("Above points successfully exported\n")
-    sys.stdout.flush()
+        above_triangles = alpha_shape(gpd.GeoSeries([Point(i) for i in above_array[:,0:2]]),0.03)
+        above_triangles_gdf = gpd.GeoDataFrame(crs=crs,geometry=[i for i in above_triangles[0] if not i.is_empty])
+        above_triangles_gdf.to_file('shp/%s_above_triangles_0.03.shp'%(outputFileName))
+        sys.stdout.write("Above points successfully triangulated and exported\n")
+        sys.stdout.flush()
 
-    above_triangles = alpha_shape(gpd.GeoSeries([Point(i) for i in above_array[:,0:2]]),0.03)
-    above_triangles_gdf = gpd.GeoDataFrame(crs=crs,geometry=[i for i in above_triangles[0] if not i.is_empty])
-    above_triangles_gdf.to_file('shp/%s_above_triangles_0.03.shp'%(outputFileName))
-    sys.stdout.write("Above points successfully triangulated and exported\n")
-    sys.stdout.flush()
+        above_polygon = cascaded_union(above_triangles_gdf.geometry)
+        above_polygon_gdf = gpd.GeoDataFrame(crs=crs,geometry=[i for i in above_polygon if not i.is_empty])
+        above_polygon_gdf.to_file('shp/%s_above_polygon.shp'%(outputFileName))
+        sys.stdout.write("Above polygon successfully exported\n")
+        sys.stdout.flush()
 
-    above_polygon = cascaded_union(above_triangles_gdf.geometry)
-    above_polygon_gdf = gpd.GeoDataFrame(crs=crs,geometry=[i for i in above_polygon if not i.is_empty])
-    above_polygon_gdf.to_file('shp/%s_above_polygon.shp'%(outputFileName))
-    sys.stdout.write("Above polygon successfully exported\n")
-    sys.stdout.flush()
-
-    # innundation = below_polygon_gdf.difference(above_polygon_gdf)
-    inundation = gpd.overlay(below_polygon_gdf,above_polygon_gdf,how='difference')
-    # inundation_gdf = gpd.GeoDataFrame(crs=crs,geometry=[i.buffer(-5,join_style=1).buffer(5,join_style=1) for i in inundation.geometry])
-    inundation_gdf = gpd.GeoDataFrame(crs=crs,geometry=[i for i in inundation.geometry])
-    inundation_gdf = inundation_gdf[inundation_gdf.is_empty==False]
-    inundation_gdf.to_file('shp/%s_inundation.shp'%(outputFileName))
+        # innundation = below_polygon_gdf.difference(above_polygon_gdf)
+        inundation = gpd.overlay(below_polygon_gdf,above_polygon_gdf,how='difference')
+        # inundation_gdf = gpd.GeoDataFrame(crs=crs,geometry=[i.buffer(-5,join_style=1).buffer(5,join_style=1) for i in inundation.geometry])
+        inundation_gdf = gpd.GeoDataFrame(crs=crs,geometry=[i for i in inundation.geometry])
+        inundation_gdf = inundation_gdf[inundation_gdf.is_empty==False]
+        inundation_gdf.to_file('shp/%s_inundation.shp'%(outputFileName))
 
     # build kml
     import fastkml
